+++ BE/src/news/news.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { NaverApiDomianService } from './naver-api-domian.service';
 import { NewsController } from './news.controller';
 import { NewsService } from './news.service';
-import { NewsRepository } from './interface/news.repository';
+import { NewsRepository } from './news.repository';
 import { News } from './news.entity';
 
 @Module({
