+++ BE/src/news/news.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectDataSource } from '@nestjs/typeorm';
 import { DataSource, Repository } from 'typeorm';
-import { News } from '../news.entity';
+import { News } from './news.entity';
 
 @Injectable()
 export class NewsRepository extends Repository<News> {
